+++ example/lib/main.dart
@@ -228,6 +228,8 @@ class _PrettyQrSettings extends StatefulWidget {
   @visibleForTesting
   static const kDefaultQrDecorationImage = PrettyQrDecorationImage(
     image: AssetImage('images/flutter.png'),
+    padding: EdgeInsets.all(8),
+    clipper: PrettyQrFlutterLogoClipper(),
     position: PrettyQrDecorationImagePosition.embedded,
   );
 
@@ -699,5 +701,28 @@ class _PrettyQrSettingsState extends State<_PrettyQrSettings> {
     imageSizeEditingController.dispose();
 
     super.dispose();
+  }
+}
+
+class PrettyQrFlutterLogoClipper implements PrettyQrClipper {
+  const PrettyQrFlutterLogoClipper();
+
+  @override
+  Path getClip(Size size) {
+    final logoPath = Path()
+      ..moveTo(0.566, 0.001)
+      ..lineTo(0.004, 0.514)
+      ..lineTo(0.525, 1.001)
+      ..lineTo(0.986, 1.001)
+      ..lineTo(0.681, 0.723)
+      ..lineTo(0.986, 0.445)
+      ..lineTo(0.577, 0.445)
+      ..lineTo(0.546, 0.417)
+      ..lineTo(1.004, 0.001)
+      ..close();
+
+    return logoPath.transform(
+      (Matrix4.identity()..scale(size.width, size.height)).storage,
+    );
   }
 }