+++ lib/src/painting/pretty_qr_shape.dart
@@ -1,5 +1,7 @@
 // ignore_for_file: prefer-static-class, delete clampDouble after min flutter version upgrade.
 
+import 'dart:ui';
+
 import 'package:meta/meta.dart';
 
 import 'package:pretty_qr_code/src/base/pretty_qr_matrix.dart';
@@ -48,14 +50,14 @@ abstract class PrettyQrShape {
   /// Paints the QR matrix on the canvas of the given painting context.
   void paint(PrettyQrPaintingContext context);
 
-  /// Same as [num.clamp] but optimized for a non-null [double].
-  @protected
-  static double clampDouble(double x, double min, double max) {
-    assert(min <= max && !max.isNaN && !min.isNaN);
-    if (x < min) return min;
-    if (x > max) return max;
-    if (x.isNaN) return max;
-    return x;
+  /// Returns a bounds of the QR code symbol painted by [paint], as an
+  /// axis-aligned Rect.
+  ///
+  /// By default, this returns a rectangle that is the same size as
+  /// the context estimated bounds. If shape paints a QR code symbol that is
+  /// roughly the same size as the context bounds, then this may be adequate.
+  Rect getBounds(PrettyQrPaintingContext context) {
+    return context.estimatedBounds;
   }
 
   /// Linearly interpolates between two [PrettyQrShape]s.
@@ -73,7 +75,7 @@ abstract class PrettyQrShape {
       return a;
     }
 
-    if (a == null) return b!.lerpFrom(null, t) ?? b;
+    if (a == null) return b?.lerpFrom(null, t) ?? b;
     if (b == null) return a.lerpTo(null, t) ?? a;
 
     if (t == 0.0) return a;
@@ -121,7 +123,7 @@ class PrettyQrCustomShape extends PrettyQrShape {
 
     final matrix = PrettyQrMatrix.masked(
       context.matrix,
-      exclude: {
+      excludeComponents: {
         if (hasFinderPattern) PrettyQrComponentType.finderPattern,
         if (hasTimingPatterns) PrettyQrComponentType.timingPattern,
         if (hasAlignmentPatterns) PrettyQrComponentType.alignmentPattern,
@@ -134,7 +136,7 @@ class PrettyQrCustomShape extends PrettyQrShape {
         context.copyWith(
           matrix: PrettyQrMatrix.masked(
             context.matrix,
-            exclude: {
+            excludeComponents: {
               for (final type in PrettyQrComponentType.values)
                 if (type != PrettyQrComponentType.finderPattern) type,
             },
@@ -148,7 +150,7 @@ class PrettyQrCustomShape extends PrettyQrShape {
         context.copyWith(
           matrix: PrettyQrMatrix.masked(
             context.matrix,
-            exclude: {
+            excludeComponents: {
               for (final type in PrettyQrComponentType.values)
                 if (type != PrettyQrComponentType.timingPattern) type,
             },
@@ -162,7 +164,7 @@ class PrettyQrCustomShape extends PrettyQrShape {
         context.copyWith(
           matrix: PrettyQrMatrix.masked(
             context.matrix,
-            exclude: {
+            excludeComponents: {
               for (final type in PrettyQrComponentType.values)
                 if (type != PrettyQrComponentType.alignmentPattern) type,
             },